requirements.txt
chainlit
langchain
langchain_community
PyPDF2
pypdf
chromadb
langchainhub
gradio
python-dotenv